[FIX] value_list_validator: do not decay to double on arithmetic

+++ test/unit/argument_parser/format_parse_validators_test.cpp
@@ -824,17 +824,17 @@ TEST(validator_test, value_list_validator_success)
 {
     // type deduction
     // --------------
-    // all arithmetic types are deduced to double in order to easily allow chaining of arithmetic validators
-    EXPECT_TRUE((std::same_as<seqan3::value_list_validator<double>,
+    // all arithmetic types are deduced to their common type in order to easily allow chaining of arithmetic validators
+    EXPECT_TRUE((std::same_as<seqan3::value_list_validator<int>,
                  decltype(seqan3::value_list_validator{1})>));
     // except char
     EXPECT_TRUE((std::same_as<seqan3::value_list_validator<char>,
                  decltype(seqan3::value_list_validator{'c'})>));
     // The same holds for a range of arithmetic types
     std::vector v{1, 2, 3};
-    EXPECT_TRUE((std::same_as<seqan3::value_list_validator<double>,
+    EXPECT_TRUE((std::same_as<seqan3::value_list_validator<int>,
                  decltype(seqan3::value_list_validator{v})>));
-    EXPECT_TRUE((std::same_as<seqan3::value_list_validator<double>,
+    EXPECT_TRUE((std::same_as<seqan3::value_list_validator<int>,
                  decltype(seqan3::value_list_validator{v | std::views::take(2)})>));
     std::vector v_char{'1', '2', '3'};
     EXPECT_TRUE((std::same_as<seqan3::value_list_validator<char>,
